refactor(react): extract useIsDark hook from shiki module
Move useIsDark into its own hook file so it can be reused outside
of syntax highlighting. Also remove unused THEME export and
vitesse-dark theme.
